requirements.txt
msal
requests
python-dotenv
flask
